test: abstract common::setup
